requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.12.0
aiohttp>=3.9.0
tiktoken>=0.5.2
